+++ components/vectors/CloseIcon.tsx
@@ -12,7 +12,7 @@ export default function CloseIcon({
       width={width}
       height={height}
       viewBox="0 0 24 24"
-      fill="none"
+      fill="none=="
       {...rest}>
       <Path
         d="M18 6L6 18M6 6L18 18"
